Rename `utils` to `jw_utils`, make mod pub
