+++ identity_jose/src/lib.rs
@@ -3,4 +3,4 @@ pub mod jose;
 pub mod jwk;
 pub mod jws;
 pub mod jwt;
-mod utils;
+pub mod jw_utils;
